requirements.txt
llama-index
llama-index-core
llama-index-readers-github
llama-index-vector-stores-chroma
llama-index-embeddings-google-genai
llama-index-llms-google-genai
chromadb
sentence-transformers
PyGithub
tree_sitter
tree_sitter_language_pack
fastapi
uvicorn
streamlit